fix: Level select button works as intended.
The problem was that I created an instance of the view controller when doing scene?.viewController = self in GameViewController.swift.
When dismissing the view controller, it still had a strong reference to it so it was not deinitialized.
When I created a new GameViewController when starting level 1 again, it wasn't able to transfer to the game over scene because the other GameViewController was conflicting with it.
Using NotificationCenter allowed me to deinit the view controller by not having a strong reference to it.